Refactored FactionClasses and Weapons
