+++ TFD_Units/Classes/CfgWeapons.hpp
@@ -1,989 +1,42 @@
 class CfgWeapons {
 
-	// CASQUETTE SOUPLE
-	class H_MilCap_blue;
-	class TFD_CasquetteGend_Depart : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Gendarmerie Departementale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteGend_Depart.paa"};
-	};
+	//CASQUES
 
-	class TFD_CasquetteGend_Mobile : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Gendarmerie Mobile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteGend_Mobile.paa"};
-	};
+	#include "Weapons\Casques\CasquetteSouple.hpp"
 
-	class TFD_CasquettePM_ASVP : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (ASVP)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM_ASVP.paa"};
-	};
+	#include "Weapons\Casques\Casquette.hpp"
 
-	class TFD_CasquettePM : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Police Municipale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM.paa"};
-	};
+	#include "Weapons\Casques\CasquetteCasqueAntiBruit.hpp"
 
-	class TFD_CasquetteAdmPen : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Administration Penitentiaire)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteAdmPen.paa"};
-	};
+	#include "Weapons\Casques\CasqueF2.hpp"
 
-	class TFD_CasquetteSP : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Sapeur-Pompier)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteSP.paa"};
-	};
+	#include "Weapons\Casques\CasqueHelicoptere.hpp"
 
-	class TFD_CasquettePM_rurale : H_MilCap_blue {
-		_generalMacro = "H_MilCap_blue";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette Souple (Police Rurale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM_rurale.paa"};
-	};
-
-	// CASQUETTE
-	class H_Cap_blk;
-	class TFD_CasquetteSecurite : H_Cap_blk {
-		_generalMacro = "H_Cap_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette (Securite)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteSecurite.paa"};
-	};
-
-	class TFD_CasquetteCadet : H_Cap_blk {
-		_generalMacro = "H_Cap_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette (Cadet)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteCadet.paa"};
-	};
-
-	class TFD_CasquettePolice : H_Cap_blk {
-		_generalMacro = "H_Cap_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette (Police)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePolice.paa"};
-	};
-
-	class TFD_CasquettePC : H_Cap_blk {
-		_generalMacro = "H_Cap_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette (Protection Civile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePC.paa"};
-	};
-
-	// CASQUETTE + CASQUE ANTI-BRUIT
-	class H_Cap_marshal;
-	class TFD_CasquetteABFormateur : H_Cap_marshal {
-		_generalMacro = "H_Cap_marshal";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casquette (Formateur)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteABFormateur.paa"};
-	};
-
-	// CASQUE F2
-	class H_HelmetB_plain_blk;
-	class TFD_CasqueF2_HDR : H_HelmetB_plain_blk {
-		_generalMacro = "H_HelmetB_plain_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque F2 (Homme du Rang)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_HDR.paa"};
-	};
-
-	class TFD_CasqueF2_SO : H_HelmetB_plain_blk {
-		_generalMacro = "H_HelmetB_plain_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque F2 (Sous-Officier)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_SO.paa"};
-	};
-
-	class TFD_CasqueF2_Off : H_HelmetB_plain_blk {
-		_generalMacro = "H_HelmetB_plain_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque F2 (Officier)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_Off.paa"};
-	};
-
-	class TFD_CasqueF2_PC : H_HelmetB_plain_blk {
-		_generalMacro = "H_HelmetB_plain_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque F2 (Protection Civile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_PC.paa"};
-	};
-
-	//CASQUE HELICOPTERE
-	class H_PilotHelmetHeli_B;
-	class TFD_CasqueHeli_GN : H_PilotHelmetHeli_B {
-		_generalMacro = "H_PilotHelmetHeli_B";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque de pilote d'helicoptere (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueHeli_GN.paa"};
-	};
-
-	class TFD_CasqueHeli_SC : H_PilotHelmetHeli_B {
-		_generalMacro = "H_PilotHelmetHeli_B";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque de pilote d'helicoptere (Securite Civile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueHeli_SC.paa"};
-	};
-
-	//CASQUE MOTO
-	class H_RacingHelmet_1_white_F;
-	class TFD_CasqueMoto_GN : H_RacingHelmet_1_white_F {
-		_generalMacro = "H_RacingHelmet_1_white_F";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Casque de moto (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueMoto_GN.paa"};
-	};
+	#include "Weapons\Casques\CasqueMoto.hpp"
 
 	// VESTES
 
-	// RECYCLEUR
-	class V_RebreatherB;
-	class TFD_RecycleurHV : V_RebreatherB {
-		_generalMacro = "V_RebreatherB";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Recycleur (Haute Visibilite)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\RecycleurHV.paa"};
-		hiddenUnderwaterSelections[] = {"hide"};
-		shownUnderwaterSelections[] = {"unhide","unhide2"};
-		hiddenUnderwaterSelectionsTextures[] = {"\TFD_Units\Data\Vest\RecycleurHV.paa","\TFD_Units\Data\Vest\RecycleurHV.paa","\A3\characters_f\data\visors_ca.paa"};
-	};
+	#include "Weapons\Vestes\Recycleur.hpp"
 
-	// GILET TACTIQUE
-	class V_TacVest_blk;
-	class TFD_GHV_Infirmier : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet Haute Visibilite (Infirmier)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Infirmier.paa"};
-	};
+	#include "Weapons\Vestes\GiletTactique.hpp"
 
-	class TFD_GHV_Medecin : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet Haute Visibilite (Medecin)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Medecin.paa"};
-	};
+	#include "Weapons\Vestes\GiletS3.hpp"
 
-	class TFD_GHV_Formateur : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet Haute Visibilite (Formateur)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Formateur.paa"};
-	};
+	#include "Weapons\Vestes\GiletAssaut.hpp"
 
-	class TFD_GiletTac_GN : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_GN.paa"};
-	};
-
-	class TFD_GiletTac_PN : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (Police)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_PN.paa"};
-	};
-
-	class TFD_GiletTac_ERIS : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (ERIS)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_ERIS.paa"};
-	};
-
-	class TFD_GiletTac_Nego : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (Negociateur)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_Nego.paa"};
-	};
-
-	class TFD_GiletTac_SC : V_TacVest_blk {
-		_generalMacro = "V_TacVest_blk";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (Deminage)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_SC.paa"};
-	};
-
-	class V_TacVest_camo;
-	class TFD_GiletTac_CE : V_TacVest_camo {
-		_generalMacro = "V_TacVest_camo";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet pare-balles (Centre Europe)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_CE.paa"};
-	};
-
-	// GILET S3
-	class V_Press_F;
-	class TFD_GiletS3_GN : V_Press_F {
-		_generalMacro = "V_Press_F";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet S3 (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_GN.paa"};
-	};
-
-	class TFD_GiletS3_PN : V_Press_F {
-		_generalMacro = "V_Press_F";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet S3 (Police)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_PN.paa"};
-	};
-
-	class TFD_GiletS3_CE : V_Press_F {
-		_generalMacro = "V_Press_F";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet S3 (Centre Europe)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_CE.paa"};
-	};
-
-	//GILET D'ASSAUT
-	class Vest_Camo_Base;
-	class VestItem;
-	class TFD_GiletAssaut_GIPN : Vest_Camo_Base {
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut (GIPN)";
-		picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
-		model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GIPN.paa"};
-		descriptionShort = "Protection ballistique III";
-		class ItemInfo: VestItem {
-			uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-			containerClass = "Supply120";
-			mass = 60;
-			armor = 20;
-			passThrough = 0.3;
-			hiddenSelections[]= {"camo"};
-		}
-	};
-
-	class TFD_GiletAssaut_Douane : Vest_Camo_Base {
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut (Douane)";
-		picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
-		model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_Douane.paa"};
-		descriptionShort = "Protection ballistique III";
-		class ItemInfo: VestItem {
-			uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-			containerClass = "Supply120";
-			mass = 60;
-			armor = 20;
-			passThrough = 0.3;
-			hiddenSelections[]= {"camo"};
-		}
-	};
-
-	class TFD_GiletAssaut_GN : Vest_Camo_Base {
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut (Gendarmerie)";
-		picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
-		model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GN.paa"};
-		descriptionShort = "Protection ballistique III";
-		class ItemInfo: VestItem {
-			uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
-			containerClass = "Supply120";
-			mass = 60;
-			armor = 20;
-			passThrough = 0.3;
-			hiddenSelections[]= {"camo"};
-		}
-	};
-
-	//GILET D'ASSAUT LOURD
-	class V_PlateCarrierIAGL_oli;
-	class TFD_GiletAssautLourd_GIPN : V_PlateCarrierIAGL_oli {
-		_generalMacro = "V_PlateCarrierIAGL_oli";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut lourd (GIPN)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo1","camo2"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GIPN.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
-	};
-
-	class TFD_GiletAssautLourd_Douane : V_PlateCarrierIAGL_oli {
-		_generalMacro = "V_PlateCarrierIAGL_oli";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut lourd (Douane)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo1","camo2"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_Douane.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
-	};
-
-	class TFD_GiletAssautLourd_GN : V_PlateCarrierIAGL_oli {
-		_generalMacro = "V_PlateCarrierIAGL_oli";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Gilet d'assaut lourd (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo1","camo2"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GN.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
-	};
+	#include "Weapons\Vestes\GiletAssautLourd.hpp"
 
 	// UNIFORME
 	class UniformItem;
 
-	// POLO
-	class U_Rangemaster;
-	class TFD_PoloGend : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Gendarmerie Nationale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloGend.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Gendarme_Depart;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
+	#include "Weapons\Uniformes\Polo.hpp"
+	
+	#include "Weapons\Uniformes\Wetsuit.hpp"
 
-	class TFD_PoloSP : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Sapeur-Pompier)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloSP.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Pompier_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
+	#include "Weapons\Uniformes\Combinaison.hpp"
 
-	class TFD_PoloSP_air : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Pompier de l'Air)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloSP_air.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PompierAir_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
+	#include "Weapons\Uniformes\Pull.hpp"
 
-	class TFD_PoloSecurite : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Securite)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloSecurite.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Vigile_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
+	#include "Weapons\Uniformes\UBACS.hpp"
 
-	class TFD_PoloPM_marine : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Police Municipale - Marine)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPM_marine.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PolicierMun_poloMarine;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloPM_blanc : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Police Municipale - Blanc)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPM_blanc.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PolicierMun_poloBlanc;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloPM_ASVP : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (ASVP)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPM_ASVP.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PolicierMun_poloASVP;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloPM_rurale : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Police Rurale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPM_rurale.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PolicierMun_poloRurale;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloAdmPen : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Administration Penitentiaire)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloAdmPen.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_SurveillantAdmPen_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloFormateur : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Formateur)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloFormateur.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Formateur_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloBenevole : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (CICR)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloCICR.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Benevole_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloDouane : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Douane)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloDouane.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Douanier_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloUIISC : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (UIISC)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloUIISC.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_SapeurSauveteur_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloPN : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Police Nationale)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPN.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_PolicierNat_Polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloCRF : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Croix Rouge Francaise)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloCRF.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_BenevoleCRF_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	class TFD_PoloPC : U_Rangemaster {
-		_generalMacro = "U_Rangemaster";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Polo (Protection Civile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PoloPC.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_BenevolePC_polo;
-            containerClass = Supply40;
-            mass = 40;
-		};
-	};
-
-	// WETSUIT
-	class U_B_Wetsuit;
-	class TFD_WetsuitHV : U_B_Wetsuit {
-		_generalMacro = "U_B_Wetsuit";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de plongee (Haute Visibilite)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\WetsuitHV.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Pompier_PL;
-            containerClass = Supply80;
-			uniformType = "Neopren";
-            mass = 60;
-		};
-	};
-
-	//Combinaison
-	class U_C_WorkerCoveralls;
-	class TFD_CombiPistard : U_C_WorkerCoveralls {
-		_generalMacro = "U_C_WorkerCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de mecanicien (Armee de l'Air)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_pistardAA.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Pistard_AA;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiDIR : U_C_WorkerCoveralls {
-		_generalMacro = "U_C_WorkerCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de patrouilleur (DIR)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_patrouilleurDIR.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Patrouilleur_DIR;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiTPB : U_C_WorkerCoveralls {
-		_generalMacro = "U_C_WorkerCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Tenue de Protection de Base";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_TPB.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Marin_TPB;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiTPBfeu : U_C_WorkerCoveralls {
-		_generalMacro = "U_C_WorkerCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Tenue de Protection de Base (Feu)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_TPBfeu.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Marin_TPBfeu;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class U_C_Scientist;
-	class TFD_CombiSAMU : U_C_Scientist {
-		_generalMacro = "U_C_Scientist";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison d'ambulancier (SAMU)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_SAMU.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Ambulancier_SAMU;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiCICR : U_C_Scientist {
-		_generalMacro = "U_C_Scientist";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de medecin (CICR)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_docCICR.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Medecin_CICR;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiCICR_sang : U_C_Scientist {
-		_generalMacro = "U_C_Scientist";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de chirurgien (CICR)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_docCICR_sang.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Chirurgien_CICR;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class U_B_HeliPilotCoveralls;
-	class TFD_CombiPilote_GN : U_B_HeliPilotCoveralls {
-		_generalMacro = "U_B_HeliPilotCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de pilote d'helicoptere (Gendarmerie)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_piloteGN.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Pilote_GN;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_CombiPilote_SC : U_B_HeliPilotCoveralls {
-		_generalMacro = "U_B_HeliPilotCoveralls";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Combinaison de pilote d'helicoptere (Securite Civile)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\Combi_piloteSC.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Pilote_SC;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	//PULL
-	class U_BG_Guerrilla_6_1;
-	class TFD_PullCE_AT : U_BG_Guerrilla_6_1 {
-		_generalMacro = "U_BG_Guerrilla_6_1";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Pull-over (Armee de Terre)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PullCE_AT.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Soldat_Pull;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	class TFD_PullMarine_AA : U_BG_Guerrilla_6_1 {
-		_generalMacro = "U_BG_Guerrilla_6_1";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Pull-over (Armee de l'Air)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\PullMarine_AA.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = TFD_Aviateur_Pull;
-            containerClass = Supply30;
-            mass = 30;
-		};
-	};
-
-	//UBACS
-	class U_B_survival_uniform;
-	class TFD_UBACSCE_Palmes : U_B_survival_uniform {
-		_generalMacro = "U_B_survival_uniform";
-        scope = 2;
-		scopeCurator = 2;
-		scopeArsenal = 2;
-        displayName="Treillis UBACS de survie (Centre Europe)";
-		author = "Heartbroken";
-		hiddenSelections[]= {"camo"};
-		hiddenSelectionsTextures[]= {"\TFD_Units\Data\Uniform\UBACS_CE.paa"};
-		class ItemInfo: UniformItem{
-            uniformModel = "-";
-            uniformClass = "TFD_Plouf_UBACS";
-            containerClass = "Supply40";
-			uniformType="Neopren";
-            mass = 60;
-		};
-	};
 };
