+++ TFD_Units/Classes/FactionClasses/TFD_AA.hpp
@@ -1,0 +1,7 @@
+class TFD_AA {
+	displayName = "Armee de l'Air";
+	priority = 5;
+	side = 1;
+	icon = "\TFD_Units\Data\Icons\Ico_AA.paa";
+	//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+};
