+++ TFD_Units/Classes/FactionClasses/TFD_AT.hpp
@@ -1,0 +1,7 @@
+class TFD_AT {
+	displayName = "Armee de Terre";
+	priority = 6;
+	side = 1;
+	icon = "\TFD_Units\Data\Icons\Ico_AT.paa";
+	//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+};
