+++ TFD_Units/Classes/FactionClasses/TFD_GN.hpp
@@ -1,0 +1,7 @@
+class TFD_GN {
+	displayName = "Gendarmerie Nationale";
+	priority = 7;
+	side = 1;
+	icon = "\TFD_Units\Data\Icons\Ico_GN.paa";
+	//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+};
