+++ TFD_Units/Classes/FactionClasses/TFD_MN.hpp
@@ -1,0 +1,7 @@
+class TFD_MN {
+	displayName = "Marine Nationale";
+	priority = 8;
+	side = 1;
+	icon = "\TFD_Units\Data\Icons\Ico_MN.paa";
+	//flag = "\a3\Data_f\Flags\flag_Altis_co.paa"; //Mettre le drapeau de la France plus tard ?
+};
