+++ TFD_Units/Classes/Weapons/Casques/CasqueF2.hpp
@@ -1,0 +1,45 @@
+// CASQUE F2
+class H_HelmetB_plain_blk;
+class TFD_CasqueF2_HDR : H_HelmetB_plain_blk {
+	_generalMacro = "H_HelmetB_plain_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque F2 (Homme du Rang)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_HDR.paa"};
+};
+
+class TFD_CasqueF2_SO : H_HelmetB_plain_blk {
+	_generalMacro = "H_HelmetB_plain_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque F2 (Sous-Officier)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_SO.paa"};
+};
+
+class TFD_CasqueF2_Off : H_HelmetB_plain_blk {
+	_generalMacro = "H_HelmetB_plain_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque F2 (Officier)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_Off.paa"};
+};
+
+class TFD_CasqueF2_PC : H_HelmetB_plain_blk {
+	_generalMacro = "H_HelmetB_plain_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque F2 (Protection Civile)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueF2_PC.paa"};
+};
