+++ TFD_Units/Classes/Weapons/Casques/CasqueHelicoptere.hpp
@@ -1,0 +1,23 @@
+//CASQUE HELICOPTERE
+class H_PilotHelmetHeli_B;
+class TFD_CasqueHeli_GN : H_PilotHelmetHeli_B {
+	_generalMacro = "H_PilotHelmetHeli_B";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque de pilote d'helicoptere (Gendarmerie)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueHeli_GN.paa"};
+};
+
+class TFD_CasqueHeli_SC : H_PilotHelmetHeli_B {
+	_generalMacro = "H_PilotHelmetHeli_B";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque de pilote d'helicoptere (Securite Civile)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueHeli_SC.paa"};
+};
