+++ TFD_Units/Classes/Weapons/Casques/CasqueMoto.hpp
@@ -1,0 +1,12 @@
+//CASQUE MOTO
+class H_RacingHelmet_1_white_F;
+class TFD_CasqueMoto_GN : H_RacingHelmet_1_white_F {
+	_generalMacro = "H_RacingHelmet_1_white_F";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casque de moto (Gendarmerie)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasqueMoto_GN.paa"};
+};
