+++ TFD_Units/Classes/Weapons/Casques/Casquette.hpp
@@ -1,0 +1,45 @@
+// CASQUETTE
+class H_Cap_blk;
+class TFD_CasquetteSecurite : H_Cap_blk {
+	_generalMacro = "H_Cap_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette (Securite)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteSecurite.paa"};
+};
+
+class TFD_CasquetteCadet : H_Cap_blk {
+	_generalMacro = "H_Cap_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette (Cadet)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteCadet.paa"};
+};
+
+class TFD_CasquettePolice : H_Cap_blk {
+	_generalMacro = "H_Cap_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette (Police)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePolice.paa"};
+};
+
+class TFD_CasquettePC : H_Cap_blk {
+	_generalMacro = "H_Cap_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette (Protection Civile)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePC.paa"};
+};
