+++ TFD_Units/Classes/Weapons/Casques/CasquetteCasqueAntiBruit.hpp
@@ -1,0 +1,12 @@
+// CASQUETTE + CASQUE ANTI-BRUIT
+class H_Cap_marshal;
+class TFD_CasquetteABFormateur : H_Cap_marshal {
+	_generalMacro = "H_Cap_marshal";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette (Formateur)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteABFormateur.paa"};
+};
