+++ TFD_Units/Classes/Weapons/Casques/CasquetteSouple.hpp
@@ -1,0 +1,78 @@
+// CASQUETTE SOUPLE
+class H_MilCap_blue;
+class TFD_CasquetteGend_Depart : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Gendarmerie Departementale)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteGend_Depart.paa"};
+};
+
+class TFD_CasquetteGend_Mobile : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Gendarmerie Mobile)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteGend_Mobile.paa"};
+};
+
+class TFD_CasquettePM_ASVP : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (ASVP)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM_ASVP.paa"};
+};
+
+class TFD_CasquettePM : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Police Municipale)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM.paa"};
+};
+
+class TFD_CasquetteAdmPen : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Administration Penitentiaire)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteAdmPen.paa"};
+};
+
+class TFD_CasquetteSP : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Sapeur-Pompier)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquetteSP.paa"};
+};
+
+class TFD_CasquettePM_rurale : H_MilCap_blue {
+	_generalMacro = "H_MilCap_blue";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Casquette Souple (Police Rurale)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Headgear\CasquettePM_rurale.paa"};
+};
