+++ TFD_Units/Classes/Weapons/Vestes/GiletAssaut.hpp
@@ -1,0 +1,65 @@
+//GILET D'ASSAUT
+class Vest_Camo_Base;
+class VestItem;
+class TFD_GiletAssaut_GIPN : Vest_Camo_Base {
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut (GIPN)";
+	picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
+	model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GIPN.paa"};
+	descriptionShort = "Protection ballistique III";
+	class ItemInfo: VestItem {
+		uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+		containerClass = "Supply120";
+		mass = 60;
+		armor = 20;
+		passThrough = 0.3;
+		hiddenSelections[]= {"camo"};
+	}
+};
+
+class TFD_GiletAssaut_Douane : Vest_Camo_Base {
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut (Douane)";
+	picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
+	model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_Douane.paa"};
+	descriptionShort = "Protection ballistique III";
+	class ItemInfo: VestItem {
+		uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+		containerClass = "Supply120";
+		mass = 60;
+		armor = 20;
+		passThrough = 0.3;
+		hiddenSelections[]= {"camo"};
+	}
+};
+
+class TFD_GiletAssaut_GN : Vest_Camo_Base {
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut (Gendarmerie)";
+	picture = "\A3\characters_f_Beta\Data\UI\icon_V_I_Vest_01_ca.paa";
+	model = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GN.paa"};
+	descriptionShort = "Protection ballistique III";
+	class ItemInfo: VestItem {
+		uniformModel = "A3\Characters_F_Beta\INDEP\equip_ia_vest01";
+		containerClass = "Supply120";
+		mass = 60;
+		armor = 20;
+		passThrough = 0.3;
+		hiddenSelections[]= {"camo"};
+	}
+};
