+++ TFD_Units/Classes/Weapons/Vestes/GiletAssautLourd.hpp
@@ -1,0 +1,34 @@
+//GILET D'ASSAUT LOURD
+class V_PlateCarrierIAGL_oli;
+class TFD_GiletAssautLourd_GIPN : V_PlateCarrierIAGL_oli {
+	_generalMacro = "V_PlateCarrierIAGL_oli";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut lourd (GIPN)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo1","camo2"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GIPN.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
+};
+
+class TFD_GiletAssautLourd_Douane : V_PlateCarrierIAGL_oli {
+	_generalMacro = "V_PlateCarrierIAGL_oli";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut lourd (Douane)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo1","camo2"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_Douane.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
+};
+
+class TFD_GiletAssautLourd_GN : V_PlateCarrierIAGL_oli {
+	_generalMacro = "V_PlateCarrierIAGL_oli";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet d'assaut lourd (Gendarmerie)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo1","camo2"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletAssaut_GN.paa","\TFD_Units\Data\Vest\GiletAssautLourd_Noir.paa"};
+};
