+++ TFD_Units/Classes/Weapons/Vestes/GiletS3.hpp
@@ -1,0 +1,34 @@
+// GILET S3
+class V_Press_F;
+class TFD_GiletS3_GN : V_Press_F {
+	_generalMacro = "V_Press_F";
+    scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+    displayName="Gilet S3 (Gendarmerie)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_GN.paa"};
+};
+
+class TFD_GiletS3_PN : V_Press_F {
+	_generalMacro = "V_Press_F";
+    scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+    displayName="Gilet S3 (Police)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_PN.paa"};
+};
+
+class TFD_GiletS3_CE : V_Press_F {
+	_generalMacro = "V_Press_F";
+    scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+    displayName="Gilet S3 (Centre Europe)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletS3_CE.paa"};
+};
