+++ TFD_Units/Classes/Weapons/Vestes/GiletTactique.hpp
@@ -1,0 +1,101 @@
+// GILET TACTIQUE
+class V_TacVest_blk;
+class TFD_GHV_Infirmier : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet Haute Visibilite (Infirmier)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Infirmier.paa"};
+};
+
+class TFD_GHV_Medecin : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet Haute Visibilite (Medecin)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Medecin.paa"};
+};
+
+class TFD_GHV_Formateur : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet Haute Visibilite (Formateur)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GHV_Formateur.paa"};
+};
+
+class TFD_GiletTac_GN : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (Gendarmerie)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_GN.paa"};
+};
+
+class TFD_GiletTac_PN : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (Police)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_PN.paa"};
+};
+
+class TFD_GiletTac_ERIS : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (ERIS)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_ERIS.paa"};
+};
+
+class TFD_GiletTac_Nego : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (Negociateur)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_Nego.paa"};
+};
+
+class TFD_GiletTac_SC : V_TacVest_blk {
+	_generalMacro = "V_TacVest_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (Deminage)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_SC.paa"};
+};
+
+class V_TacVest_camo;
+class TFD_GiletTac_CE : V_TacVest_camo {
+	_generalMacro = "V_TacVest_camo";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName="Gilet pare-balles (Centre Europe)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Vest\GiletTac_CE.paa"};
+};
